收货地址策略权限

+++ app/Http/Controllers/UserAddressController.php
@@ -37,11 +37,13 @@ class UserAddressController extends Controller
 
     public function edit(UserAddress $user_address)
     {
+        $this->authorize('update', $user_address);
         return view('user_address.create_and_edit', ['address' => $user_address]);
     }
 
     public function update(UserAddress $user_address, UserAddressRequest $request)
     {
+        $this->authorize('update', $user_address);
         $user_address->update($request->only([
             'province',
             'city',
@@ -57,6 +59,8 @@ class UserAddressController extends Controller
 
     public function destroy(UserAddress $user_address)
     {
+
+        $this->authorize('update', $user_address);
         $user_address->delete();
         return redirect()->route('user_address.index');
     }
